Updated README with performance notes.
The rendering system with components is 4X faster than my previous
rendering system. Also, thinking about meshes and locii as the
essential (but separate) fundamental components of rendering lets
me create other systems for modifying the locus and mesh of any
entity. So a ribbon will be a line/ribbon component and path expansion
system that modifies a given mesh component.