Erase SpriteSystem elements safely.
Will do a speed comparison now with the erase happening to make
sure vector iteration is still faster (since we need to do a removal
every update to make it safe). We could optimize the removal a bit
by storing the dead ones and looping through that list to erase things.