Removed PhysicsSystem from generic system list.
Gets too specific to the application at hand to really live in
this base level. May reappear as PlatformerPhysicsSystem or some
other useful base implementation in the future.
In the meantime, suBox2d is improving with my understanding of
Erin Catto's library and is a good library to build a PhysicsSystem
on top of.